Extract ClaimForm from new-claim page
